% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospdeath.R
\name{hosp_load_scenario_sim}
\alias{hosp_load_scenario_sim}
\title{Data loading utility function for this package.}
\usage{
hosp_load_scenario_sim(
  prefix,
  run_id,
  sim_id,
  keep_compartments = NULL,
  time_filter_low = -Inf,
  time_filter_high = Inf,
  geoid_len = 0,
  padding_char = "0",
  use_parquet = FALSE
)
}
\arguments{
\item{sim_id}{integer Which file of the files in `scenario_dir`.}

\item{keep_compartments}{character Optional names of compartments to keep (and throw away others).  Default is to keep all compartments}

\item{time_filter_low}{lubridate::date time_filter_low Only keep rows that have time after this}

\item{time_filter_high}{lubridate::date time_filter_low Only keep rows that have time before this}

\item{padding_char}{If geoids need to be increased in length, which character should be used to add length}

\item{use_parquet}{Whether to save to parquet files rather than csvs}

\item{scenario_dir}{The directory to load scenarios from}

\item{geod_len}{The minimum length each geoid should be}
}
\description{
Load the csv associated with a single scenario
}
