% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hospdeath.R
\name{build_hospdeath_par}
\alias{build_hospdeath_par}
\title{Build a set of sampled hospitalizations, deaths, and recoveries
from the incident infection data from the simulation model.}
\usage{
build_hospdeath_par(
  p_hosp,
  p_death,
  p_ICU,
  p_vent,
  dscenario_name,
  time_hosp_pars = c(1.23, 0.79),
  time_ICU_pars = c(log(10.5), log((10.5 - 7)/1.35)),
  time_vent_pars = c(log(10.5), log((10.5 - 8)/1.35)),
  time_hosp_death_pars = c(log(11.25), log(1.15)),
  time_disch_pars = c(log(11.5), log(1.22)),
  time_ICUdur_pars = c(log(17.46), log(4.044)),
  time_ventdur_pars = log(17),
  cores = 8,
  use_parquet = FALSE,
  start_sim = 1,
  num_sims = -1,
  in_run_id = covidcommon::run_id(),
  in_prefix = covidcommon::create_prefix(dscenario_name, in_run_id, trailing_separator
    = "/"),
  out_run_id = covidcommon::run_id(),
  out_prefix = covidcommon::create_prefix(dscenario_name, out_run_id,
    trailing_separator = "/")
)
}
\arguments{
\item{p_hosp}{probability of hospitalization, among infections}

\item{p_death}{probability of death, among infections (hospitalization is required for death)}

\item{p_ICU}{probability of needing the ICU among hospitalized patients}

\item{p_vent}{probability of needing a ventilator among ICU patients}

\item{dscenario_name}{The name of the death scenario we are analyzing here (e.g., "highdeath", "meddeath", etc.)}

\item{time_hosp_pars}{parameters for time from onset to hospitalization distribution}

\item{time_ICU_pars}{parameters for time from hospitalization to ICU}

\item{time_vent_pars}{parameters for time from ICU to time on ventilator}

\item{time_hosp_death_pars}{parameters for time from hospitalization to death distribution}

\item{time_disch_pars}{parameters for time from hospitalization to discharge parameters}

\item{time_ICUdur_pars}{parameetrs for time of ICU duration}

\item{cores}{The number of CPU cores to run this model on in parallel}

\item{use_parquet}{Whether to save to parquet files rather than csvs}

\item{root_out_dir}{Path to the directory to write the outputs of this analysis}
}
\description{
Build a set of sampled hospitalizations, deaths, and recoveries
from the incident infection data from the simulation model.
}
